// NutriMais/src/app/resourcers/models/requestConsulta.ts
export class requestConsulta{
    user: number;
    nutricionista: number;
    peso_atual: number;
    quantidade_refeicoes: number;
    quantidade_refeicoes_almejadas: number;
    refeicoes: string
    valor_disponivel: number;
    rotina: string;
    alimentos_favoritos: string;
    alimentos_nao_desejaveis: string;
    alergias: string;
    exercicio_fisico: string;
    restricao_medica: string;
    consumo_agua: number;
    objetivo: string
}